feat(licensing): implement license key system and update build configuration
- Add LICENSE_SYSTEM_GUIDE.txt documentation for license key generation
- Add scripts/generate_license_key.bat batch script for Windows license key generation
- Update LicenseManager.cpp and LicenseManager.h to implement license validation logic
- Refactor AuricHalo.jucer project configuration with cleaner formatting and updated plugin settings
- Remove MIDI input requirement (pluginWantsMidiIn changed from 1 to 0)
- Remove standalone application build configuration (delete StandaloneApp.cpp)
- Reorganize JUCE library includes by removing headless variants and consolidating ARA/LV2 support
- Update CMakeLists.txt build configuration to align with new licensing system
- Improve LevelMeter.cpp UI component implementation
- Update AuricHaloEditor.cpp and AuricHaloProcessor.cpp to integrate license manager
- Simplify JUCE project file attributes by removing redundant resource declarations
- This change enables license key protection for the plugin while maintaining cross-platform compatibility

+++ source/AuricHaloEditor.cpp
@@ -319,10 +319,10 @@ void AuricHaloEditor::paint(juce::Graphics& g)
     g.setColour(juce::Colour(0xffd4af37));
     g.setFont(juce::FontOptions("Arial", labelSize, juce::Font::bold));
     g.drawText("HALO ENGINE", 
-               (float)btnBounds.getX() - 20, 
-               labelY, 
-               (float)btnBounds.getWidth() + 40, 
-               15.0f * textScale, 
+               static_cast<int>((float)btnBounds.getX() - 20), 
+               static_cast<int>(labelY), 
+               static_cast<int>((float)btnBounds.getWidth() + 40), 
+               static_cast<int>(15.0f * textScale), 
                juce::Justification::centred);
 }
